requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
python-dotenv
PyPDF2
tqdm